Fixed: 1. Integration Table automation: corrected the result slot to extraction-only access so external automation can retrieve completed outputs without inserting into the result slot. 2. Requester comparator: signal strength now measures fulfillment against the configured request amount, making fully satisfied requests consistently report full strength. 3. Item transfers: made ItemTransactorHelper source-first and rollback-safe, preventing duplication and server-tick failures when inventories change between simulation and execution. 4. Robot tools and searches: removed duplicate tool durability consumption and made transient searches restart safely after reload while preserving already-reserved goto targets. 5. Mining progress: widened machine and Stripes breaking progress to 64-bit values, preventing overflow on expensive blocks and fixing high-progress crack animation. 6. Kinesis persistence: MJ pipe current and queued energy buffers now survive chunk unloads and restarts and request periodic saves while populated. 7. Stripes extensions: deferred requests now validate the exact originating pipe and refund stale requests instead of crashing, deleting items, or acting on replacement pipes. 8. Water Gel: removed obsolete scheduled ticks so growth remains exclusively random-tick driven and follows randomTickSpeed. 9. Regression coverage: added transactional item-transfer and MJ-buffer persistence GameTests plus dedicated P1 CI guards across all supported targets.

+++ source-platforms/neoforge/src/main/java/buildcraft/transport/pipe/behaviour/PipeBehaviourStripes.java
@@ -63,7 +63,7 @@ public class PipeBehaviourStripes extends PipeBehaviour implements StripesOutput
 
     @Nullable
     public Direction direction = null;
-    private int progress;
+    private long progress;
 
     public PipeBehaviourStripes(IPipe pipe) {
         super(pipe);
